🎨 style: 修复 Clippy 警告和代码格式化
- 移除不必要的类型转换 (as u64)
- 为未使用的辅助函数添加 #[allow(dead_code)] 属性
- 修复未使用的导入和变量
- 应用 rustfmt 代码格式化

修复内容：
- src/instruction/raydium_amm_v4.rs: 移除 2 处不必要的类型转换
- src/instruction/utils/raydium_clmm/tick_array_bitmap.rs: 添加 allow 属性
- src/liquidity/cpmm.rs: 修复未使用的变量
- src/utils/calc/raydium_amm_v4.rs: 移除未使用的导入

+++ src/instruction/raydium_amm_v4.rs
@@ -19,6 +19,7 @@ use solana_sdk::{
 ///
 /// WSOL 固定使用 TOKEN_PROGRAM
 /// USDC 自动检测（支持 Token-2022）
+#[allow(dead_code)]
 fn get_input_token_program(is_wsol: bool) -> &'static solana_sdk::pubkey::Pubkey {
     if is_wsol {
         &crate::constants::TOKEN_PROGRAM
@@ -237,7 +238,7 @@ impl InstructionBuilder for RaydiumAmmV4InstructionBuilder {
         // 如果 max_amount_in 刚好等于计算值，扣除费用后可能无法得到足够的输出
         let (param1, param2) = if params.fixed_output_amount.is_some() {
             // exact_out: 添加滑点缓冲到 max_amount_in
-            let slippage_bps = params.slippage_basis_points.unwrap_or(DEFAULT_SLIPPAGE) as u64;
+            let slippage_bps = params.slippage_basis_points.unwrap_or(DEFAULT_SLIPPAGE);
             let max_amount_in = amount_in.saturating_add((amount_in * slippage_bps / 10000).max(1));
             (max_amount_in, minimum_amount_out)
         } else {
@@ -444,7 +445,7 @@ impl InstructionBuilder for RaydiumAmmV4InstructionBuilder {
         // 在 exact_out 模式下，需要添加滑点缓冲到 max_amount_in
         let (param1, param2) = if params.fixed_output_amount.is_some() {
             // exact_out: 添加滑点缓冲到 max_amount_in
-            let slippage_bps = params.slippage_basis_points.unwrap_or(DEFAULT_SLIPPAGE) as u64;
+            let slippage_bps = params.slippage_basis_points.unwrap_or(DEFAULT_SLIPPAGE);
             let max_amount_in = amount_in.saturating_add((amount_in * slippage_bps / 10000).max(1));
             (max_amount_in, minimum_amount_out)
         } else {
